Shift formatTime outside of StorageExplorerStore to StorageExplroer.utils
