+++ apps/studio/components/to-be-cleaned/Storage/StorageExplorer/StorageExplorer.utils.tsx
@@ -11,6 +11,20 @@ export const copyPathToFolder = (openedFolders: StorageItem[], item: StorageItem
   toast.success(`Copied path to folder "${item.name}"`)
 }
 
+export const formatTime = (seconds: number) => {
+  const days = Math.floor(seconds / (24 * 3600))
+  seconds %= 24 * 3600
+  const hours = Math.floor(seconds / 3600)
+  seconds %= 3600
+  const minutes = Math.floor(seconds / 60)
+  seconds = Math.floor(seconds % 60)
+
+  if (days > 0) return `${days}d `
+  if (hours > 0) return `${hours}h `
+  if (minutes > 0) return `${minutes}m `
+  return `${seconds}s`
+}
+
 export const downloadFile = async ({
   projectRef,
   bucketId,
